refactor(phase-3B-partial): extract viz websocket + ui_helpers from server.py [no behavior change]
Phase 3B Step 1 / viz/REFACTORING.md §2 Step 1 — first wave of the
viz/server.py split.  Two clusters move to sibling modules; the
class-based HTTP handler ``_Handler.do_GET`` chain stays in
``server.py`` for now (1000+ lines, slated for PR-3B-1's later step).

New modules:
- ``ari/viz/websocket.py`` — ``_ws_handler`` async WebSocket function
  (was ``server.py:48``).  ``server.py`` re-exports it via
  ``from .websocket import _ws_handler`` so ``ws_serve(_ws_handler, ...)``
  in ``main()`` is unchanged.
- ``ari/viz/ui_helpers.py`` — ``_extract_goal_from_md`` /
  ``_build_experiment_detail_config`` / ``_collect_resource_metrics``
  + the ``_REDACT_KEYS`` constant (was ``server.py:62-214``).
  ``server.py`` re-exports each name; downstream callers and the
  remaining route handlers continue to look them up off the
  ``server`` module.

Test fixture update:
- ``test_settings_roundtrip._srv()`` now concatenates ``ui_helpers.py``
  + ``websocket.py`` + ``server.py`` (in that order so
  ``_REDACT_KEYS`` is defined ahead of the import line that mentions
  it) — keeps the existing source-text grep checks working without
  weakening them.

pytest ari-core/tests/ -q (excluding live integrations):
1971 passed, 15 skipped.

Phase 3B continues in follow-up commits with the route-table split
(PR-3B-1) + ``api_state.py`` 7-way split (PR-3B-2); both stay in
``server.py`` / ``api_state.py`` for now.
